% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{bind_cps}
\alias{bind_cps}
\title{bind list of cps files together}
\usage{
bind_cps(x, version_check)
}
\arguments{
\item{x}{list of cps files}

\item{version_check}{whether to check verisons}
}
\description{
bind list of cps files together
}
