% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{cps_metadata}
\alias{cps_metadata}
\alias{cps_version}
\alias{cps_citation}
\alias{assert_cps_version}
\title{Retrieve metadata from CPS extract}
\usage{
cps_version(x)

cps_citation(x)

assert_cps_version(x, version)
}
\arguments{
\item{x}{EPI CPS extract generated from \code{load_cps()} functions}

\item{version}{String version number}
}
\value{
\code{cps_version} and \code{cps_citation} return version or citation strings.

\code{assert_cps_version} returns an error when the provided version is incorrect.
}
\description{
Retrieve metadata from CPS extract
}
\examples{
cps_org <- load_org_sample(2023:2025)
cps_citation(cps_org)
cps_version(cps_org)
}
