% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{crosstab}
\alias{crosstab}
\title{crosstab: cross-tabulations}
\usage{
crosstab(df, x, y = NULL, w = NULL)
}
\arguments{
\item{df}{a data frame}

\item{x}{first variable}

\item{y}{second variable}

\item{w}{weight}
}
\value{
a tibble
}
\description{
crosstab: cross-tabulations
}
\examples{
crosstab(mtcars, cyl, gear)
}
