% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_cps}
\alias{download_cps}
\title{Download CPS data}
\usage{
download_cps(sample, extracts_dir = NULL, overwrite = FALSE)
}
\arguments{
\item{sample}{CPS sample ("org", "basic", "may")}

\item{extracts_dir}{directory where EPI extracts should be placed}

\item{overwrite}{when TRUE, overwrite data}
}
\value{
downloaded files
}
\description{
More description
}
\examples{
\dontrun{
download_cps(sample = "march", extracts_dir = "/data/cps")
}
}
