% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{read_single_year}
\alias{read_single_year}
\title{Read single year of feather CPS}
\usage{
read_single_year(
  year,
  sample,
  variables = NULL,
  extracts_dir,
  version_check = TRUE
)
}
\arguments{
\item{year}{year of data}

\item{sample}{CPS sample ("basic", "may", "org")}

\item{variables}{variables to keep}

\item{extracts_dir}{directory where EPI extracts are}

\item{version_check}{when TRUE, confirm separate monthly files have same version}
}
\description{
Read single year of feather CPS
}
