% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{valid_sample_name}
\alias{valid_sample_name}
\title{check sample name validity}
\usage{
valid_sample_name(x)
}
\arguments{
\item{x}{sample name}
}
\value{
valid, lowercase sample name
}
\description{
check sample name validity
}
