% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{crosstab}
\alias{crosstab}
\title{crosstab: cross-tabulations}
\usage{
crosstab(data, ..., w = NULL, col = FALSE, row = FALSE)
}
\arguments{
\item{data}{a data frame}

\item{...}{one or two variables, for a one- or two-way cross-tabulation}

\item{w}{weight}

\item{col}{column percentages in a two-way cross-tabulation}

\item{row}{row percentages in a two-way cross-tabulation}
}
\value{
a tibble
}
\description{
crosstab: cross-tabulations
}
\examples{
crosstab(mtcars, cyl)
crosstab(mtcars, cyl, gear)
crosstab(mtcars, cyl, gear, w = mpg, col = TRUE)
}
