% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_cps}
\alias{load_cps}
\alias{load_basic}
\alias{load_may}
\alias{load_org}
\title{Load a selection of EPI CPS extracts}
\usage{
load_cps(.sample, .years, ..., .extracts_dir = NULL, .version_check = TRUE)

load_basic(.years, ..., .extracts_dir = NULL, .version_check = TRUE)

load_may(.years, ..., .extracts_dir = NULL, .version_check = TRUE)

load_org(.years, ..., .extracts_dir = NULL, .version_check = TRUE)
}
\arguments{
\item{.sample}{CPS sample ("org", "basic", "march", "may")}

\item{.years}{years of CPS data (integers)}

\item{...}{tidy selection of variables to keep}

\item{.extracts_dir}{directory where EPI extracts are}

\item{.version_check}{when TRUE, confirm data are same version}
}
\value{
A tibble of CPS microdata
}
\description{
Select years and variables from the EPI CPS microdata extracts. These data
must first be downloaded using \code{download_cps()} or from
\url{https://microdata.epi.org}.
}
\details{
All columns are selected if \code{...} is missing.

\code{.extracts_dir} is required, but if NULL it will look for the environment variables

\if{html}{\out{<div class="sourceCode">}}\preformatted{EPIEXTRACTS_CPSBASIC_DIR
EPIEXTRACTS_CPSORG_DIR
EPIEXTRACTS_CPSMAY_DIR
}\if{html}{\out{</div>}}

which could be set in your .Renviron, for example.
}
\section{Functions}{
\itemize{
\item \code{load_cps()}: base function group

\item \code{load_basic()}: Load CPS Basic Monthly files

\item \code{load_may()}: Load CPS May files

\item \code{load_org()}: Load CPS ORG files

}}
\examples{
\dontrun{
# Load all of the 2019-2020 CPS Basic columns
load_basic(2019:2020)

# Load a selection of 2010-2019 CPS ORG columns:
load_org(2010:2019, year, month, orgwgt, female, wage)

# These are equivalent:
load_org(2019:2020)
load_cps("org", 2019:2020)
}
}
