% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_cps}
\alias{load_cps}
\title{Load CPS data}
\usage{
load_cps(
  years,
  sample,
  variables = NULL,
  extracts_dir = NULL,
  version_check = TRUE
)
}
\arguments{
\item{years}{years of CPS data (integers)}

\item{sample}{CPS sample ("org", "basic", "march", "may")}

\item{variables}{variables to keep}

\item{extracts_dir}{directory where EPI extracts are}

\item{version_check}{when TRUE, confirm data are same version}
}
\value{
a tibble of CPS microdata
}
\description{
More description
}
\examples{
\dontrun{
load_cps(years = 2018:2019, sample = "org")
}
}
